+++ parallel/src/sgpp/parallel/pde/application/HeatEquationSolverMPI.cpp
@@ -38,9 +38,9 @@ HeatEquationSolverMPI::~HeatEquationSolverMPI() {
   }
 }
 
-void HeatEquationSolverMPI::constructGrid(sgpp::base::BoundingBox& BoundingBox, int level) {
+void HeatEquationSolverMPI::constructGrid(sgpp::base::BoundingBox& BoundingBox, size_t level) {
   this->dim = BoundingBox.getDimensions();
-  this->levels = level;
+  this->levels = static_cast<int>(level);
 
   this->myGrid = new sgpp::base::LinearBoundaryGrid(BoundingBox);
